When adding test data, hack GroupCount timestamp
Because the MessageTracker in briar initializes the GroupCount timestamp
to the current time when the contact is being added, we cannot have
realistic test data in the database where the GroupCount timestamp
matches that of the last message in the chat history. This is a dirty
workaround to make this possible.