Don't use text PointerIcon on buttons in text fields
Thanks to the research of Paul, I could quickly find the right modifier
to change the mouse appearance upon hovering buttons.

For most buttons we stay with the default PointerIcon, just inside text
fields it showed the text PointerIcon. With this commit the default
PointerIcon is shown consistently across all buttons.

In a later MR we could switch all buttons to show the hand icon, but I
wouldn't do it because later Compose versions might do this
automatically and the APIs used here are experimental.