Don't make AuthorView clickable in blogs
because we only have the WRAPPED_POST which is in the Blog Group of the re-poster. Therefore, it isn't easy to get the Blog GroupId of the re-posted author. It may even be that we don't have a group, because the re-posted author isn't a contact or their blog wasn't shared with us.